created an abstract class BookBase
Implemented an abstract class BookBase to create a base layout for the book views, this will allow for coordinating the same values across the views. This will also allow for fragments to placed below the existing base layout so buttons can be interchanged more easily while reducing the duplicate code in each class.